--- a/Jenkinsfile
+++ b/Jenkinsfile
@@ -171,7 +171,7 @@ pipeline {
                 stage('Build and unit test with coverage') {
                     steps {
                         timeout(time: 20, unit: 'MINUTES') {
-                            sh './gradlew build coverage --scan'
+                            sh './gradlew build coverage'
                         }
                     }
                 }
@@ -267,13 +267,6 @@ pipeline {
             }
         }
 
-        stage ('Javascript Test and Coverage') {
-            when { expression { changeClass in ['full', 'api-catalog'] } }
-            steps {
-                sh './gradlew :api-catalog-ui:startMockedBackend &'
-                sh './gradlew :api-catalog-ui:javaScriptCoverage'
-            }
-        }
 
         stage ('Codecov') {
             when { expression { changeClass in ['full', 'api-catalog'] } }