--- a/api-catalog-ui/build.gradle
+++ b/api-catalog-ui/build.gradle
@@ -11,7 +11,7 @@ node {
 
     // If true, it will download node using above parameters.
     // If false, it will try to use globally installed node.
-    download = true
+    download = false
 
     // Set the work directory for unpacking node
     workDir = file("${project.projectDir}/tools/nodejs")