--- a/api-catalog-ui/build.gradle
+++ b/api-catalog-ui/build.gradle
@@ -49,6 +49,7 @@ task runTests(type: NpmTask) {
 task pnpmInstall(type: Exec) {
     def operativeSystem = System.properties['os.name']
     inputs.file("package.json")
+    outputs.dir('frontend/node_modules')
     if (operativeSystem.contains("Windows"))
         commandLine "pnpm.cmd", "--prefix", "${project.projectDir}/frontend/", "install"
     else