--- a/api-catalog-ui/frontend/package.json
+++ b/api-catalog-ui/frontend/package.json
@@ -76,7 +76,7 @@
     "devDependencies": {
         "babel-cli": "^6.26.0",
         "babel-eslint": "10.0.3",
-        "babel-jest": "^24.8.0",
+        "babel-jest": "^24.9.0",
         "babel-preset-flow": "6.23.0",
         "babel-preset-react-app": "3.1.2",
         "body-parser": "1.18.3",