--- a/api-catalog-ui/frontend/package.json
+++ b/api-catalog-ui/frontend/package.json
@@ -50,8 +50,6 @@
         "test:int": "react-scripts test --env=jsdom integration-tests/*",
         "test:e2e": "react-scripts test --env=jsdom e2e/*",
         "test:e2e:ci": "cross-env REACT_APP_CATALOG_URL_TEST=${API_CATALOG_BASE_URL} REACT_APP_CATALOG_USERNAME=${MF_USERID} REACT_APP_CATALOG_PASSWORD=${MF_PASSWORD} CI=true react-scripts test --env=jsdom src/e2e/*.test.jsx --reporters=default --reporters=jest-junit",
-        "galen": "galen",
-        "test:galen": "galen test src\\responsive-tests\\tests/ --htmlreport report",
         "coverage": "npm test -- --coverage --watchAll=false",
         "cy:open": "cypress open",
         "cy:e2e": "cypress run --spec cypress/integration/e2e/**/*.test.js --env baseURL=https://ca3x.ca.com:10010/,username=${MF_USERID},password=${MF_PASSWORD}",