--- a/build.gradle
+++ b/build.gradle
@@ -188,3 +188,9 @@ afterReleaseBuild.dependsOn publishAllVersions
 //-----------Release part end
 
 
+if (hasProperty('buildScan')) {
+    buildScan {
+        termsOfServiceUrl = 'https://gradle.com/terms-of-service'
+        termsOfServiceAgree = 'yes'
+    }
+}