--- a/gradle.properties
+++ b/gradle.properties
@@ -25,4 +25,4 @@ nodejsVersion=8.11.1
 projectRoot=${project.projectDir}
 
 org.gradle.daemon=false
-org.gradle.parallel=true
+org.gradle.parallel=false