--- a/gradle.properties
+++ b/gradle.properties
@@ -24,5 +24,5 @@ defaultSpringBootCloudVersion=2.0.0.RELEASE
 nodejsVersion=8.11.1
 projectRoot=${project.projectDir}
 
-org.gradle.daemon=false
-org.gradle.parallel=false
+org.gradle.daemon=true
+org.gradle.parallel=true