… api-catalog-ui/frontend/pnpm-lock.yaml
… api-catalog-ui/frontend/src/responsive-tests/galen-tests.MD
… api-catalog-ui/frontend/src/responsive-tests/specs/tiles.gspec
… api-catalog-ui/frontend/src/responsive-tests/tests/dashboard.test.js
… api-catalog-ui/frontend/src/responsive-tests/tests/pages/dashboardPage.jsx