## Problem

During the build of APIML, a big amount of the time is spent for installing npm dependencies and for building the React-based UI module.

## Solution

A possible solution to speed up the build is to use the [pnpm](https://pnpm.js.org) package manager instead of `npm` during the installation phase (`npmInstall` task in `Gradle`)

Link with some additional benchmark [here](https://github.com/pnpm/pnpm) comparing `yarn`, `npm` and `pnpm`.
The main advantages of using `pnpm` are that it links files from its cache rather than copying them and does not use the flat structure that `npm` has moved to in recent years.
`pnpm` supports the same commands as `npm`. Operations not directly implemented by `pnpm` are passed through to `npm` to execute.

Some benchmark has been executed in local environment.

BENCHMARK of dependencies installation for API Catalog UI:

|            | npm install | pnpm install |
|------------|-------------|--------------|
| Windows 10 | 2min 43sec  | 50sec        |
| Mac/OS     | 1min 10sec  | 33sec        |

BENCHMARK of API Catalog UI build (gradle api-cagalog-ui build task):

|            | npm build  | pnpm build |
|------------|------------|------------|
| Windows 10 | 6min 12sec | 3min 46sec |
| Mac/OS     | 2min 15sec        | 2min 8sec  |

**Note:** There are not noticeable improvements in Jenkins (additional investigation in the future to understand the reason)

**Note:** `package-lock.json` will not be part of versioning anymore. Instead, `pnpm` uses `pnpm-lock.yaml` which should be included on GitHub.

## Improvement of Jenkins Pipeline

1. Removed redundant javascript coverage step from Jenkins.
2. Defined output directory for the `pnpmInstall` gradle task. In this way, during the `sonar` step of Jenkins, Gradle will automatically skip this task, saving time.

## Outcome

1. Significative improvements of performance for the npm build process have been achieved in Windows and Mac/OS environment. 
2. Future investigation can be done in order to improve the build speed in Jenkins.
3. Clean up and improvement of some step in Jenkins have been done.
4. The size of node_modules is decreased thanks to Pnpm, of around 24%.
## How to test

1. Run `npm add -g pnpm` to install it.
2. Run `gradle api-catalog-ui:build` task